+++ pkg/querier/tenantfederation/merge_queryable_test.go
@@ -12,13 +12,14 @@ import (
 	"reflect"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/go-kit/log"
 	"github.com/grafana/dskit/tenant"
 	"github.com/grafana/dskit/user"
 	"github.com/opentracing/opentracing-go"
 	"github.com/opentracing/opentracing-go/mocktracer"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/testutil"
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/storage"
@@ -64,12 +65,6 @@ func (m *mockTenantQueryableWithFilter) Querier(_, _ int64) (storage.Querier, er
 		queryErrByTenant: m.queryErrByTenant,
 	}
 	return q, nil
-}
-
-// UseQueryable implements the querier.QueryableWithFilter interface.
-// It ensures the mockTenantQueryableWithFilter storage.Queryable is always used.
-func (m *mockTenantQueryableWithFilter) UseQueryable(_ time.Time, _, _ int64) bool {
-	return true
 }
 
 type mockTenantQuerier struct {
@@ -149,8 +144,8 @@ func (m *mockSeriesSet) Warnings() annotations.Annotations {
 
 // Select implements the storage.Querier interface.
 func (m mockTenantQuerier) Select(ctx context.Context, _ bool, _ *storage.SelectHints, matchers ...*labels.Matcher) storage.SeriesSet {
-	log, _ := spanlogger.NewWithLogger(ctx, m.logger, "mockTenantQuerier.select")
-	defer log.Span.Finish()
+	logger, _ := spanlogger.NewWithLogger(ctx, m.logger, "mockTenantQuerier.select")
+	defer logger.Span.Finish()
 	var matrix model.Matrix
 
 	tenantID, err := tenant.TenantID(ctx)
@@ -261,9 +256,10 @@ type mergeQueryableScenario struct {
 	doNotByPassSingleQuerier bool
 }
 
-func (s *mergeQueryableScenario) init(t *testing.T) (context.Context, storage.Querier) {
+func (s *mergeQueryableScenario) init(t *testing.T) (context.Context, prometheus.Gatherer, storage.Querier) {
 	// initialize with default tenant label
-	q := NewQueryable(&s.queryable, !s.doNotByPassSingleQuerier, defaultConcurrency, log.NewNopLogger())
+	reg := prometheus.NewPedanticRegistry()
+	q := NewQueryable(&s.queryable, !s.doNotByPassSingleQuerier, defaultConcurrency, reg, log.NewNopLogger())
 
 	// inject tenants into context
 	ctx := context.Background()
@@ -274,7 +270,7 @@ func (s *mergeQueryableScenario) init(t *testing.T) (context.Context, storage.Qu
 	// retrieve querier
 	querier, err := q.Querier(mint, maxt)
 	require.NoError(t, err)
-	return ctx, querier
+	return ctx, reg, querier
 }
 
 // selectTestCase is the inputs and expected outputs of a call to Select.
@@ -291,6 +287,8 @@ type selectTestCase struct {
 	expectedWarnings []string
 	// expectedQueryErr is the error expected when querying.
 	expectedQueryErr error
+	// expectedMetrics is the expected metrics emitted by the merge queryable
+	expectedMetrics string
 }
 
 // selectScenario tests a call to Select over a range of test cases in a specific scenario.
@@ -311,6 +309,8 @@ type labelNamesTestCase struct {
 	expectedWarnings []string
 	// expectedQueryErr is the error expected when querying.
 	expectedQueryErr error
+	// expectedMetrics is the expected metrics emitted by the merge queryable
+	expectedMetrics string
 }
 
 // labelNamesScenario tests a call to LabelNames in a specific scenario.
@@ -333,6 +333,8 @@ type labelValuesTestCase struct {
 	expectedWarnings []string
 	// expectedQueryErr is the error expected when querying.
 	expectedQueryErr error
+	// expectedMetrics is the expected metrics emitted by the merge queryable
+	expectedMetrics string
 }
 
 // labelValuesScenario tests a call to LabelValues over a range of test cases in a specific scenario.
@@ -343,10 +345,11 @@ type labelValuesScenario struct {
 
 func TestMergeQueryable_Querier(t *testing.T) {
 	t.Run("querying without a tenant specified should error", func(t *testing.T) {
+		reg := prometheus.NewPedanticRegistry()
 		queryable := &mockTenantQueryableWithFilter{
 			logger: log.NewNopLogger(),
 		}
-		qable := NewQueryable(queryable, false /* bypassWithSingleID */, defaultConcurrency, log.NewNopLogger())
+		qable := NewQueryable(queryable, false /* bypassWithSingleID */, defaultConcurrency, reg, log.NewNopLogger())
 		q, err := qable.Querier(mint, maxt)
 		require.NoError(t, err)
 
@@ -404,12 +407,37 @@ var (
 			},
 		},
 	}
+
+	expectSingleTenantsMetrics = `
+# HELP cortex_querier_federation_sample_tenants_queried Number of tenants queried for a single standard query.
+# TYPE cortex_querier_federation_sample_tenants_queried histogram
+cortex_querier_federation_sample_tenants_queried_bucket{le="1"} 1
+cortex_querier_federation_sample_tenants_queried_bucket{le="2"} 1
+cortex_querier_federation_sample_tenants_queried_bucket{le="4"} 1
+cortex_querier_federation_sample_tenants_queried_bucket{le="8"} 1
+cortex_querier_federation_sample_tenants_queried_bucket{le="16"} 1
+cortex_querier_federation_sample_tenants_queried_bucket{le="32"} 1
+cortex_querier_federation_sample_tenants_queried_bucket{le="+Inf"} 1
+cortex_querier_federation_sample_tenants_queried_sum 1
+cortex_querier_federation_sample_tenants_queried_count 1
+`
+
+	expectThreeTenantsMetrics = `
+# HELP cortex_querier_federation_sample_tenants_queried Number of tenants queried for a single standard query.
+# TYPE cortex_querier_federation_sample_tenants_queried histogram
+cortex_querier_federation_sample_tenants_queried_bucket{le="1"} 0
+cortex_querier_federation_sample_tenants_queried_bucket{le="2"} 0
+cortex_querier_federation_sample_tenants_queried_bucket{le="4"} 1
+cortex_querier_federation_sample_tenants_queried_bucket{le="8"} 1
+cortex_querier_federation_sample_tenants_queried_bucket{le="16"} 1
+cortex_querier_federation_sample_tenants_queried_bucket{le="32"} 1
+cortex_querier_federation_sample_tenants_queried_bucket{le="+Inf"} 1
+cortex_querier_federation_sample_tenants_queried_sum 3
+cortex_querier_federation_sample_tenants_queried_count 1
+`
 )
 
 func TestMergeQueryable_Select(t *testing.T) {
-	// Set a multi tenant resolver.
-	tenant.WithDefaultResolver(tenant.NewMultiResolver())
-
 	for _, scenario := range []selectScenario{
 		{
 			mergeQueryableScenario: threeTenantsScenario,
@@ -417,21 +445,25 @@ func TestMergeQueryable_Select(t *testing.T) {
 				{
 					name:                "should return all series when no matchers are provided",
 					expectedSeriesCount: 6,
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 				{
 					name:                "should return only series for team-a and team-c tenants when there is a not-equals matcher for the team-b tenant",
 					matchers:            []*labels.Matcher{{Name: defaultTenantLabel, Value: "team-b", Type: labels.MatchNotEqual}},
 					expectedSeriesCount: 4,
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 				{
 					name:                "should return only series for team-b when there is an equals matcher for the team-b tenant",
 					matchers:            []*labels.Matcher{{Name: defaultTenantLabel, Value: "team-b", Type: labels.MatchEqual}},
 					expectedSeriesCount: 2,
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 				{
 					name:                "should return one series for each tenant when there is an equals matcher for the host1 instance",
 					matchers:            []*labels.Matcher{{Name: "instance", Value: "host1", Type: labels.MatchEqual}},
 					expectedSeriesCount: 3,
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 			},
 		},
@@ -449,37 +481,44 @@ func TestMergeQueryable_Select(t *testing.T) {
 						labels.FromStrings("__tenant_id__", "team-c", "instance", "host1", "original___tenant_id__", "original-value", "tenant-team-c", "static"),
 						labels.FromStrings("__tenant_id__", "team-c", "instance", "host2.team-c", "original___tenant_id__", "original-value"),
 					},
+					expectedMetrics: expectThreeTenantsMetrics,
 				},
 				{
 					name:                "should return only series for team-a and team-c tenants when there is with not-equals matcher for the team-b tenant",
 					matchers:            []*labels.Matcher{{Name: defaultTenantLabel, Value: "team-b", Type: labels.MatchNotEqual}},
 					expectedSeriesCount: 4,
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 				{
 					name: "should return no series where there are conflicting tenant matchers",
 					matchers: []*labels.Matcher{
 						{Name: defaultTenantLabel, Value: "team-a", Type: labels.MatchEqual}, {Name: defaultTenantLabel, Value: "team-c", Type: labels.MatchEqual}},
 					expectedSeriesCount: 0,
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 				{
 					name:                "should return only series for team-b when there is an equals matcher for team-b tenant",
 					matchers:            []*labels.Matcher{{Name: defaultTenantLabel, Value: "team-b", Type: labels.MatchEqual}},
 					expectedSeriesCount: 2,
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 				{
 					name:                "should return all series when there is an equals matcher for the original value of __tenant_id__ using the revised tenant label",
 					matchers:            []*labels.Matcher{{Name: originalDefaultTenantLabel, Value: "original-value", Type: labels.MatchEqual}},
 					expectedSeriesCount: 6,
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 				{
 					name:                "should return all series when there is a regexp matcher for the original value of __tenant_id__ using the revised tenant label",
 					matchers:            []*labels.Matcher{labels.MustNewMatcher(labels.MatchRegexp, originalDefaultTenantLabel, "original-value")},
 					expectedSeriesCount: 6,
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 				{
 					name:                "should return no series when there is a not-equals matcher for the original value of __tenant_id__ using the revised tenant label",
 					matchers:            []*labels.Matcher{{Name: originalDefaultTenantLabel, Value: "original-value", Type: labels.MatchNotEqual}},
 					expectedSeriesCount: 0,
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 			},
 		},
@@ -492,6 +531,7 @@ func TestMergeQueryable_Select(t *testing.T) {
 					`warning querying tenant_id team-c: out of office`,
 				},
 				expectedSeriesCount: 6,
+				expectedMetrics:     expectThreeTenantsMetrics,
 			},
 			}},
 		{
@@ -499,19 +539,21 @@ func TestMergeQueryable_Select(t *testing.T) {
 			selectTestCases: []selectTestCase{{
 				name:             "should return any error encountered with any tenant",
 				expectedQueryErr: errors.New("error querying tenant_id team-b: failure xyz"),
+				expectedMetrics:  expectThreeTenantsMetrics,
 			}},
 		},
 	} {
 		t.Run(scenario.name, func(t *testing.T) {
 			for _, tc := range scenario.selectTestCases {
 				t.Run(tc.name, func(t *testing.T) {
-					ctx, querier := scenario.init(t)
+					ctx, reg, querier := scenario.init(t)
 					seriesSet := querier.Select(ctx, true, &storage.SelectHints{Start: mint, End: maxt}, tc.matchers...)
 
 					if tc.expectedQueryErr != nil {
 						require.EqualError(t, seriesSet.Err(), tc.expectedQueryErr.Error())
 					} else {
 						require.NoError(t, seriesSet.Err())
+						assert.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(tc.expectedMetrics), "cortex_querier_federation_sample_tenants_queried"))
 						assertEqualWarnings(t, tc.expectedWarnings, seriesSet.Warnings())
 					}
 
@@ -534,15 +576,13 @@ func TestMergeQueryable_Select(t *testing.T) {
 }
 
 func TestMergeQueryable_LabelNames(t *testing.T) {
-	// set a multi tenant resolver
-	tenant.WithDefaultResolver(tenant.NewMultiResolver())
-
 	for _, scenario := range []labelNamesScenario{
 		{
 			mergeQueryableScenario: singleTenantScenario,
 			labelNamesTestCase: labelNamesTestCase{
 				name:               "should not return the __tenant_id__ label as the MergeQueryable has been bypassed",
 				expectedLabelNames: []string{"instance", "tenant-team-a"},
+				expectedMetrics:    expectSingleTenantsMetrics,
 			},
 		},
 		{
@@ -551,6 +591,7 @@ func TestMergeQueryable_LabelNames(t *testing.T) {
 				name:               "should not return the __tenant_id__ label as the MergeQueryable has been bypassed with matchers",
 				matchers:           []*labels.Matcher{labels.MustNewMatcher(labels.MatchRegexp, seriesWithLabelNames, "bar|foo")},
 				expectedLabelNames: []string{"bar", "foo", "instance", "tenant-team-a"},
+				expectedMetrics:    expectSingleTenantsMetrics,
 			},
 		},
 		{
@@ -558,6 +599,7 @@ func TestMergeQueryable_LabelNames(t *testing.T) {
 			labelNamesTestCase: labelNamesTestCase{
 				name:               "should return the __tenant_id__ label as the MergeQueryable has not been bypassed",
 				expectedLabelNames: []string{defaultTenantLabel, "instance", "tenant-team-a"},
+				expectedMetrics:    expectSingleTenantsMetrics,
 			},
 		},
 		{
@@ -565,6 +607,7 @@ func TestMergeQueryable_LabelNames(t *testing.T) {
 			labelNamesTestCase: labelNamesTestCase{
 				name:               "should return the __tenant_id__ label and all tenant team labels",
 				expectedLabelNames: []string{defaultTenantLabel, "instance", "tenant-team-a", "tenant-team-b", "tenant-team-c"},
+				expectedMetrics:    expectThreeTenantsMetrics,
 			},
 		},
 		{
@@ -572,6 +615,7 @@ func TestMergeQueryable_LabelNames(t *testing.T) {
 			labelNamesTestCase: labelNamesTestCase{
 				name:               "should return  the __tenant_id__ label and all tenant team labels, and the __original_tenant_id__ label",
 				expectedLabelNames: []string{defaultTenantLabel, "instance", originalDefaultTenantLabel, "tenant-team-a", "tenant-team-b", "tenant-team-c"},
+				expectedMetrics:    expectThreeTenantsMetrics,
 			},
 		},
 		{
@@ -583,6 +627,7 @@ func TestMergeQueryable_LabelNames(t *testing.T) {
 					`warning querying tenant_id team-b: don't like them`,
 					`warning querying tenant_id team-c: out of office`,
 				},
+				expectedMetrics: expectThreeTenantsMetrics,
 			},
 		},
 		{
@@ -590,6 +635,7 @@ func TestMergeQueryable_LabelNames(t *testing.T) {
 			labelNamesTestCase: labelNamesTestCase{
 				name:             "should return any error encountered with any tenant",
 				expectedQueryErr: errors.New("error querying tenant_id team-b: failure xyz"),
+				expectedMetrics:  expectThreeTenantsMetrics,
 			},
 		},
 		{
@@ -602,6 +648,7 @@ func TestMergeQueryable_LabelNames(t *testing.T) {
 					`warning querying tenant_id team-b: don't like them`,
 					`warning querying tenant_id team-c: out of office`,
 				},
+				expectedMetrics: expectThreeTenantsMetrics,
 			},
 		},
 		{
@@ -617,6 +664,7 @@ func TestMergeQueryable_LabelNames(t *testing.T) {
 				expectedWarnings: []string{
 					`warning querying tenant_id team-b: don't like them`,
 				},
+				expectedMetrics: expectThreeTenantsMetrics,
 			},
 		},
 		{
@@ -632,6 +680,7 @@ func TestMergeQueryable_LabelNames(t *testing.T) {
 					`warning querying tenant_id team-b: don't like them`,
 					`warning querying tenant_id team-c: out of office`,
 				},
+				expectedMetrics: expectThreeTenantsMetrics,
 			},
 		},
 		{
@@ -646,18 +695,20 @@ func TestMergeQueryable_LabelNames(t *testing.T) {
 				expectedWarnings: []string{
 					`warning querying tenant_id team-b: don't like them`,
 				},
+				expectedMetrics: expectThreeTenantsMetrics,
 			},
 		},
 	} {
 		t.Run(scenario.mergeQueryableScenario.name, func(t *testing.T) {
 			t.Run(scenario.labelNamesTestCase.name, func(t *testing.T) {
-				ctx, querier := scenario.init(t)
+				ctx, reg, querier := scenario.init(t)
 				labelNames, warnings, err := querier.LabelNames(ctx, scenario.labelNamesTestCase.matchers...)
 				if scenario.labelNamesTestCase.expectedQueryErr != nil {
 					require.EqualError(t, err, scenario.labelNamesTestCase.expectedQueryErr.Error())
 				} else {
 					require.NoError(t, err)
 					assert.Equal(t, scenario.labelNamesTestCase.expectedLabelNames, labelNames)
+					assert.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(scenario.labelNamesTestCase.expectedMetrics), "cortex_querier_federation_sample_tenants_queried"))
 					assertEqualWarnings(t, scenario.labelNamesTestCase.expectedWarnings, warnings)
 				}
 			})
@@ -666,9 +717,6 @@ func TestMergeQueryable_LabelNames(t *testing.T) {
 }
 
 func TestMergeQueryable_LabelValues(t *testing.T) {
-	// set a multi tenant resolver
-	tenant.WithDefaultResolver(tenant.NewMultiResolver())
-
 	for _, scenario := range []labelValuesScenario{
 		{
 			mergeQueryableScenario: singleTenantScenario,
@@ -677,11 +725,13 @@ func TestMergeQueryable_LabelValues(t *testing.T) {
 					name:                "should return all label values for instance when no matchers are provided",
 					labelName:           "instance",
 					expectedLabelValues: []string{"host1", "host2.team-a"},
+					expectedMetrics:     expectSingleTenantsMetrics,
 				},
 				{
 					name:                "should return no tenant values for the __tenant_id__ label as the MergeQueryable has been bypassed",
 					labelName:           defaultTenantLabel,
 					expectedLabelValues: nil,
+					expectedMetrics:     expectSingleTenantsMetrics,
 				},
 			},
 		},
@@ -692,11 +742,13 @@ func TestMergeQueryable_LabelValues(t *testing.T) {
 					name:                "should return all label values for instance when no matchers are provided",
 					labelName:           "instance",
 					expectedLabelValues: []string{"host1", "host2.team-a"},
+					expectedMetrics:     expectSingleTenantsMetrics,
 				},
 				{
 					name:                "should return a tenant team value for the __tenant_id__ label as the MergeQueryable has not been bypassed",
 					labelName:           defaultTenantLabel,
 					expectedLabelValues: []string{"team-a"},
+					expectedMetrics:     expectSingleTenantsMetrics,
 				},
 			},
 		},
@@ -707,6 +759,7 @@ func TestMergeQueryable_LabelValues(t *testing.T) {
 					name:                "should return all label values for instance when no matchers are provided",
 					labelName:           "instance",
 					expectedLabelValues: []string{"host1", "host2.team-a", "host2.team-b", "host2.team-c"},
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 				{
 					name:      "should propagate non-tenant matchers to downstream queriers",
@@ -719,6 +772,7 @@ func TestMergeQueryable_LabelValues(t *testing.T) {
 						"warning querying tenant_id team-b: " + mockMatchersNotImplemented,
 						"warning querying tenant_id team-c: " + mockMatchersNotImplemented,
 					},
+					expectedMetrics: expectThreeTenantsMetrics,
 				},
 				{
 					name: "should return no values for the instance label when there are conflicting tenant matchers",
@@ -728,29 +782,34 @@ func TestMergeQueryable_LabelValues(t *testing.T) {
 					},
 					labelName:           "instance",
 					expectedLabelValues: []string{},
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 				{
 					name:                "should only query tenant-b when there is an equals matcher for team-b tenant",
 					matchers:            []*labels.Matcher{{Name: defaultTenantLabel, Value: "team-b", Type: labels.MatchEqual}},
 					labelName:           "instance",
 					expectedLabelValues: []string{"host1", "host2.team-b"},
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 				{
 					name:                "should return all tenant team values for the __tenant_id__ label when no matchers are provided",
 					labelName:           defaultTenantLabel,
 					expectedLabelValues: []string{"team-a", "team-b", "team-c"},
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 				{
 					name:                "should return only label values for team-a and team-c tenants when there is a not-equals matcher for team-b tenant",
 					labelName:           defaultTenantLabel,
 					matchers:            []*labels.Matcher{{Name: defaultTenantLabel, Value: "team-b", Type: labels.MatchNotEqual}},
 					expectedLabelValues: []string{"team-a", "team-c"},
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 				{
 					name:                "should return only label values for team-b tenant when there is an equals matcher for team-b tenant",
 					labelName:           defaultTenantLabel,
 					matchers:            []*labels.Matcher{{Name: defaultTenantLabel, Value: "team-b", Type: labels.MatchEqual}},
 					expectedLabelValues: []string{"team-b"},
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 			},
 		},
@@ -761,16 +820,19 @@ func TestMergeQueryable_LabelValues(t *testing.T) {
 					name:                "should return all label values for instance when no matchers are provided",
 					labelName:           "instance",
 					expectedLabelValues: []string{"host1", "host2.team-a", "host2.team-b", "host2.team-c"},
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 				{
 					name:                "should return all tenant values for __tenant_id__ label name",
 					labelName:           defaultTenantLabel,
 					expectedLabelValues: []string{"team-a", "team-b", "team-c"},
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 				{
 					name:                "should return the original value for the revised tenant label name when no matchers are provided",
 					labelName:           originalDefaultTenantLabel,
 					expectedLabelValues: []string{"original-value"},
+					expectedMetrics:     expectThreeTenantsMetrics,
 				},
 				{
 					name:                "should return the original value for the revised tenant label name with matchers",
@@ -782,6 +844,7 @@ func TestMergeQueryable_LabelValues(t *testing.T) {
 						"warning querying tenant_id team-b: " + mockMatchersNotImplemented,
 						"warning querying tenant_id team-c: " + mockMatchersNotImplemented,
 					},
+					expectedMetrics: expectThreeTenantsMetrics,
 				},
 			},
 		},
@@ -795,6 +858,7 @@ func TestMergeQueryable_LabelValues(t *testing.T) {
 					`warning querying tenant_id team-b: don't like them`,
 					`warning querying tenant_id team-c: out of office`,
 				},
+				expectedMetrics: expectThreeTenantsMetrics,
 			}},
 		},
 		{
@@ -803,6 +867,7 @@ func TestMergeQueryable_LabelValues(t *testing.T) {
 				name:                "should not return warnings as the underlying queryables are not queried in requests for the __tenant_id__ label",
 				labelName:           defaultTenantLabel,
 				expectedLabelValues: []string{"team-a", "team-b", "team-c"},
+				expectedMetrics:     expectThreeTenantsMetrics,
 			}},
 		},
 		{
@@ -811,6 +876,7 @@ func TestMergeQueryable_LabelValues(t *testing.T) {
 				name:             "should return any error encountered with any tenant",
 				labelName:        "instance",
 				expectedQueryErr: errors.New("error querying tenant_id team-b: failure xyz"),
+				expectedMetrics:  expectThreeTenantsMetrics,
 			}},
 		},
 		{
@@ -819,19 +885,21 @@ func TestMergeQueryable_LabelValues(t *testing.T) {
 				name:                "should not return errors as the underlying queryables are not queried in requests for the __tenant_id__ label",
 				labelName:           defaultTenantLabel,
 				expectedLabelValues: []string{"team-a", "team-b", "team-c"},
+				expectedMetrics:     expectThreeTenantsMetrics,
 			}},
 		},
 	} {
 		t.Run(scenario.name, func(t *testing.T) {
 			for _, tc := range scenario.labelValuesTestCases {
 				t.Run(tc.name, func(t *testing.T) {
-					ctx, querier := scenario.init(t)
+					ctx, reg, querier := scenario.init(t)
 					actLabelValues, warnings, err := querier.LabelValues(ctx, tc.labelName, tc.matchers...)
 					if tc.expectedQueryErr != nil {
 						require.EqualError(t, err, tc.expectedQueryErr.Error())
 					} else {
 						require.NoError(t, err)
 						assert.Equal(t, tc.expectedLabelValues, actLabelValues, fmt.Sprintf("unexpected values for label '%s'", tc.labelName))
+						assert.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(tc.expectedMetrics), "cortex_querier_federation_sample_tenants_queried"))
 						assertEqualWarnings(t, tc.expectedWarnings, warnings)
 					}
 				})
@@ -893,12 +961,11 @@ func TestSetLabelsRetainExisting(t *testing.T) {
 func TestTracingMergeQueryable(t *testing.T) {
 	mockTracer := mocktracer.New()
 	opentracing.SetGlobalTracer(mockTracer)
+	reg := prometheus.NewPedanticRegistry()
 	ctx := user.InjectOrgID(context.Background(), "team-a|team-b")
 
-	// set a multi tenant resolver
-	tenant.WithDefaultResolver(tenant.NewMultiResolver())
 	filter := mockTenantQueryableWithFilter{}
-	q := NewQueryable(&filter, false, defaultConcurrency, log.NewNopLogger())
+	q := NewQueryable(&filter, false, defaultConcurrency, reg, log.NewNopLogger())
 	// retrieve querier if set
 	querier, err := q.Querier(mint, maxt)
 	require.NoError(t, err)